Move functions to tools.py
