--- services/extrainaipe.py
@@ -1,2 +1,0 @@
-def extrai_naipe(carta:str) -> str:
-    return carta[-1]
