--- services/extraivalor.py
@@ -1,7 +1,0 @@
-def extrai_valor(carta:str) -> str:
-    
-    if (len(carta) == 3):
-        return carta[:2]
-    
-    return carta[0]
-
